Reorganize directives
