+++ cli/commands/core/index.ts
@@ -1,5 +1,6 @@
 import yargs, { Argv } from "yargs";
 import { readCommand } from "./read";
+import { addModuleCommand, executePHOUpdateCommand, updateModuleCeilingCommand } from "./write";
 
 export const coreCommand = {
     command: "core",
@@ -7,6 +8,9 @@ export const coreCommand = {
     builder: (yargs: Argv): yargs.Argv => {
       return yargs
       .command(readCommand)
+      .command(addModuleCommand)
+      .command(updateModuleCeilingCommand)
+      .command(executePHOUpdateCommand)
     },
     handler: (): void => {
       yargs.showHelp();
