+++ cli/commands/core/write/index.ts
