+++ cli/commands/core/write/updateCeiling.ts
@@ -4,20 +4,19 @@ import { loadEnv } from "../../../env";
 import { execute, generateForgeCommand, generateSignature } from "../../deploy";
 import { verifyModule } from "../../../helpers";
 import { CLIArgs, CLIEnvironment, CommandParams } from "../../../types";
+import { ethers } from "ethers";
+
 
 const buildHelp = () => {
-    let help = "To execute module ceiling -> modules execute-ceiling [moduleId]";
-    return help;
-  };
-  
-export const executeCeilingUpdate = async (
-  cli: CLIEnvironment,
-  cliArgs: CLIArgs,
-): Promise<void> => {
-  let { c: networkId, moduleId } = cliArgs;
+  let help = "To update module ceiling -> core update-ceiling [moduleId] [ceiling]";
+  return help;
+};
+
+export const updateCeiling = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
+  let { c: networkId, moduleId, ceiling } = cliArgs;
 
   if (!verifyModule(networkId, moduleId)) return;
-
+  let ceilingD18 = ethers.utils.parseUnits(ceiling, 18);
   let sig: string = await generateSignature([
     {
       type: "string",
@@ -27,10 +26,14 @@ export const executeCeilingUpdate = async (
       type: "address",
       value: moduleId,
     },
+    {
+      type: "uint256",
+      value: ceilingD18,
+    },
   ]);
 
   let commandParams: CommandParams = {
-    contractName: "UpdateExecuteCeilingUpdate",
+    contractName: "UpdateModulePHOCeiling",
     forkUrl: cli.providerUrl,
     privateKey: cli.wallet.privateKey,
     sig,
@@ -39,16 +42,16 @@ export const executeCeilingUpdate = async (
   let forgeCommand = generateForgeCommand(commandParams);
   await execute(forgeCommand);
 
-  logger.info(`Successfully updated ceiling for module ${moduleId}`);
+  logger.info(`Successfully updated PHO ceiling for module ${moduleId}`);
 };
 
-export const executePHOUpdateCommand = {
-  command: "execute-ceiling [moduleId]",
-  describe: "Executes ceiling update for a module",
+export const updateModuleCeilingCommand = {
+  command: "update-ceiling [moduleId] [ceiling]",
+  describe: "Updates PHO ceiling for a given module",
   builder: (yargs: Argv): yargs.Argv => {
     return yargs.usage(buildHelp());
   },
   handler: async (argv: CLIArgs): Promise<void> => {
-    return executeCeilingUpdate(await loadEnv(argv), argv);
+    return updateCeiling(await loadEnv(argv), argv);
   },
 };
