+++ cli/commands/index.ts
@@ -1,11 +1,9 @@
 import { coreCommand } from "./core";
-import { modulesCommand } from "./modules";
 import { deployCommand } from "./deploy";
 import { evmCommand } from "./evm";
 
 export { 
     coreCommand,
-    modulesCommand,
     deployCommand,
     evmCommand
 }
